Correction of GTM connection initialization at backend start
When Coordinator <-> Coordinator connections were created due to a DDL
or a utility query, backend coordinator initialized connections to GTM and Pooler
on uselessly.
This was increasing slightly the number of backends on GTM, possibly
causing a crash due to a too high number of threads opened on GTM.

A sudden crash of GTM may have huge impact on data visibility or
consistency, so this bug was perhaps the origin of multiple problems
seen during performance tests.

It is thought that this bug is present in XC since the implementation of
DDL synchronization, done in August/September 2010.