In following cases, we push the aggregates and group by clauses to the datanode and collect the results at coordinator 1. The query is single step 2. it does not have having clause, order by clause, windowing clause, distinct clause 3. the target list has only aggregates (no expressions involving aggregates) and/or grouping clause expressions
If these conditions are met, then following steps are taken,
1. Create RemoteQuery node with GROUP BY clause, and clauses corresponding to
the targetlist, Scan nodes in the lefttree of the Agg node etc.
2. The target list of Agg node is modified such that the aggregate nodes accept
the corresponding aggregate nodes of the RemoteQuery as inputs. That enables
feeding the transition results from the datanodes to the aggregates on
coordinator.
3. The grouping columns in Agg node are modified to the corresponding columns in
the target list of the RemoteQuery node.

During execution, the transition results from the datanodes are grouped
according to the group by clause and "collected" at coordinator and finalized at
the end for each group. The patch adds the machinary for group wise collections
and finalization of collected results.

This work will be refined later to lift the restrictions as much.