This patch adds a system in XC to cancel a running query, and flush network buffers of any results data nodes might have sent before cancelling the query. This was required to fix certain issues where coordinator encounters an error while processing rows from data nodes and quits row processing. It then issues a new query and finds an old row description in the network buffer. This can and was crashing the server.
To cancel a query a new pooler command 'h' is added.
This command is sent to the pooler by the coordinator and the pooler issues PQcancel to the respective data nodes.
Cancel request is sent every time coordinator raises an error of level more than ERROR.

This commit fixes bug 3306801