Support for TEMPORARY sequences
A temporary sequence is created only on local Coordinator.
It doesn't need any interaction with GTM as it only lives during the
session session is up.
Sequence algorithm for nextval, currval and lastval has been changed in
accordance to temporary sequences.

Moreover, this commit cleans up a part of the code used by XC in utility.c
to have a more effective lookup on objects when looking for the node
types utility queries should be sent to.

COMMENTS are changed to be executed only on Coordinators.
Comments on temporary objects need some special handling but this is
not targetted yet.