Support for temporary TABLE/VIEW
Temporary tables and views are supported so as it uses the same
interface as PostgreSQL.
For Postgres-XC, it is possible for users to create un the cluster
temporary tables that are replicated or distributed the same way
as permanent tables.
As those objects are session-limited, creation is made automatically on
the Datanodes at the moment of query launch like normal DDL.
Views are created only on local Coordinators.
Support for ON COMMIT DROP/PRESERVE ROWS/DELETE ROWS is guarranted the
same way as PostgreSQL.
Support for INDEX, DROP, RENAME is also guarranted like normal PostgreSQL.

One of the main changes is the addition of a DISCARD ALL clean up command
run in pooler for pooler connections used during session to Datanodes when
session finishes. This command is run the same way as RESET ALL for SET commands
at the end of a session.

A new regression test case called xc_temp has been added to check JOINs between
temporary and persistent tables for the types types of distribution supported by
Postgres-XC: REPLICATION, HASH (equivalent to MODULO), ROUND ROBIN.