Fix for bug 3389327: query deparse issue
Origin of this bug was XC planner trying to identify the keyword "FROM"
in a query during reconstruction of an ORDER BY query.

The trick here is to replace "FROM" in the aliases generated by query deparsing
by "_FROM_" to avoid having query reconstruction look for aliases instead of
real FROM clauses.

The new algorithm in place in XC planner is able to rebuild query aliases
to more consistent names even in multiple places.

This fixes the regression test date.