Addition of pgxc_prepared_xacts
This new system view is based on a plpgsql function that uses
EXECUTE DIRECT to scan GIDs of prepared transactions on each node
and gather results back to the client.
pgxc_prepared_xacts returns a distinct list of GIDs.

pg_prepared_xacts is kept the same as the original one in PostgreSQL.
As a consequence, schema scanning is not used anymore when looking
for prepared transaction list.
Now users should use pgxc_prepared_xacts to recover the list related
to the whole cluster. pg_prepared_xacts can only be used on a local node.