Support for make check
"make check" is a compilation command that can be run to check
if code has been correctly compiled in the environment.
For Postgres-XC, this is run on 2 Coordinators and 2 Datanodes
that are installed on local machine in the temporary repository
chosen.

Coordinator 1 has data folder data_co1.
Coordinator 2 has data folder data_co2.
Datanode 1 has data folder data_dn1.
Datanode 2 has data folder data_dn2.

The entry port of for regression client that can be customized with
the option "--port" is Coordinator 1, and all the regressions are made
from this one to the other remote nodes.
A configuration with 2 Coordinators is chosen to check that DDL queries
are correctly run.

"make check" is basically supported by there are still uses with XC itself
when running a parallel schedule in regressions.