Fix for bug 3410137: CREATE TABLE ... LIKE with temp table
Treat correctly table inheritance dependencies between temporary
and permanent objects by insuring temporary object creation on
all the nodes. This enforces creation of temporary tables not only
on all the datanodes but also on coordinators so it increases
interactions between nodes but this is necessary to keep synchronized
catalogs for tables.

A test case in regressions is added regarding this particular behavior.