Fix for an assertion failure when running regressions
The origin of the assertion error was double:
1) DISCARD queries send directly from the pooler did not include
transaction IDs necessary also on remote nodes. This finishes
by having a transaction committing with a local transaction ID,
not a global one.
2) When a connection is cut between 2 Postgres-XC nodes and session
on remote node hold temporary objects, a transaction ID was consumed
to clean up the temporary objects with a callback API. This transaction
finished using a local transaction ID, where it should have used a global
one.

The use of local transaction IDs for commit does not protect commit
consistency at all and may create issues with transaction IDs committed
twice in the worse case. Sometimes it also created issues with
assertions in 2PC commit.

Some simplifications in pooler are added, and to insure that temporary
objects are correctly deleted from the system at each session ending,
connections holding temporary objects are not pull back to pull but simply
destroyed.
There may be some more performant way to do that, but it may involve to send
on all the nodes incolved in session a "DISCARD TEMP" query at least with a
transaction ID with a snapshot, and I am not sure this is worth doing at
session ending on Coordinator. So for the time being only cut connection.
Connections that used only session parameters or normal objects except
temporary ones are put back normally to pooler.