Fix a server crash is query deparsing. PostgreSQL has introduced a new API for query deparse since 9.1, but that works only for init-ed plans. We were using that API on a Plan node where as it expects a PlanState node to work on. This causes segfault, especially on a 32-bit machine, but the code was wrong anyways.
We create dummy PlanState nodes for the purpose of deparse. Since the
use is quite limited right now, it works. But this is just a work-around
and in the long term, we should either add appropriate API for query
deparsing at planning time or move the logic of generating query string at
the execution time. The second option looks more promising, but the remote
join reduction logic may need some rework.