Fix for bug 3434781: join involving replicated table
The following sequence was failing:
create table aa (a int) distribute by hash(a);
create table bb (a int) distribute by replication;
select * from aa join bb using (a) where a=1;

This last query failed with an unidentified table.
The origin of this error was in XC planner where JOIN rte was
evaluated as a table for location data.