Fix for bug 3431570: crash with Hadoop and sqoop
This crash was happening at the execution of a RemoteQuery node
when trying to allocate memory for fresh handles taken from pool.
In this case allocation was made in TopTransactionContext.
However, in the case of sqoop, which is a Hadoop module creating an
interface with a db backend, it happened that TopTransactionContext
was NULL, leading to a crash of the node.

This commit switches the memory context to CurrentMemoryContext
and contains fixes for possible memory leaks related to barrier and
COPY.

Fix from Pavan Deolasee