Support for data distribution among a subset of datanodes
CREATE TABLE has been extended with the following clause:
CREATE TABLE ...
[ TO ( GROUP groupname | NODE nodename [, ... ] ) ]

This clause allows to distribute data among a subset of nodes
listed by a node list, or a group alias.
Node groups can be defined with CREATE NODE GROUP.

The base structure for this support was added with commit
56a90674444df1464c8e7012c6113efd7f9bc7db, but check of mapping of
subsets of node list was still missing for the management of join
push down and materialization evaluation in planner.