Forbid access on Coordinator for inconsistent connection data
An additional check is made when session starts up to see if remote
node information is consistent between pool and catalogs.
In case it is not, non-superusers are not allowed to have an access
to the cluster as this could result in creation of inconsistent data.
A superuser is authorized to connect however he receives a warning
message to inform that remote node information inconsistency has
to be solved.