Support CTAS for node subsets and add documentation
CTAS is supports node subsets the same way as CREATE TABLE.
This commit includes fixes for warnings of related to CTAS
and its documentation.

A fix is made with a problem found with the special fix for
round robin tables. It happened that other table's sublist were
not used anymore, leading to inconsistencies when using INSERT
SELECT on a subset of nodes.