Add support for UPDATE/DELETE WHERE CURRENT OF
This patch adds support for UPDATE/DELETE using WHERE CURRENT OF
In order to identify rows with in a data node ctid is used
and in order to find the data node itself data node name is used.
Both ctid and data node name are added to the target list
of the cursor query and when the user issues an UPDATE/DELETE
using WHERE CURRENT OF, we identify the data node using node name
and row using ctid.

One limitation of current implementation is that it does not
support successive UPDATEs or an UPDATE followed by a DELETE
because the first UPDATE changes the ctid.