Improve expression push-down in INSERT queries
In XC planner, if non-shippable expression is found in the target
list of INSERT values, redirect query through standard planner.
This fix allows to run INSERT queries with non-foreign expression
as stable and volatile functions among other things.

This commit has consequences on the regression test largeobject
but we do not intend to support that for the time being.