Fix for Bug id:3425212: Failed to build on Ubuntu 11.10
Ubuntu 11 no longer has the libpthread.so in the usual locations like /lib or
/usr/lib. We have hardcoded -lpthread linker option used in gtm module.
The fix is to instead use PTHREAD_CFLAGS macro that is defined in
src/Makefile.global. This macro has the correct pthread options already. In
this particular issue, the -pthread (not -lpthread) option was the one that
was needed to resolve the compile errors.

The backend postgres executable also needs this option in case of XC, because
it uses functions from src/interfaces/libpq/fe-secure.c which in turn uses
pthread calls.