Support for remote UPDATE planning in standard planning
This commit focuses on support of UPDATE queries using non-pushable
expressions in SET and WHERE clauses.
The UPDATE query is generated based on the output of the SELECT query
generated by create_remotequery_plan to select the correct tuples to
update remotelly with a top plan that was not generated before.

Current implementation has a restriction though:
on replicated tables all the columns cannot be updated at once if the
query is not entirely pushable. Postgres-XC planner is in charge of
analyzing that.
For distributed tables, the distribution column cannot be updated
so this restriction is not directly applied but this should be taken
into account once a tuple relocation mechanism is added.