Support for EXPLAIN with remote DML planning
When running an explain query with DML that need Coordinator to
be generated, additional remote query plans in ModifyTable are
printed. Those plans are created by create_remoteinsert_plan,
create_remoteupdate_plan and create_remotedelete_plan in createplan.c

This commit also includes a fix for remote DML plans related to the scan
relation ID used. The relation ID was used but the correct way is to
use its Index when referring to the table list in Scan plan.