Support for TABLESPACE
TABLESPACE are re-enabled without having any modifications in its grammar.
A user defining a location for tablespace needs to be sure that this
location exists on all the nodes of the cluster.
A suffix with node name is added on the PG_XXX folder created in tablespace
to support tablespace in case several nodes are running on the same node.
New tablespace folder format is PG_XXX_NodeName. As node name is unique
and constant in the cluster, this insures that no conflict will appear
at the creation level.

For the time begin, DBA has no way to refine location of tablespace for a
single node level but this will be added later with extra support of CREATE/DROP
TABLESPACE through EXECUTE DIRECT.