Stamp HEAD as 1.0devel.
This stamp modifies package name and adds a version number for Postgres-XC
by keeping the version of PostgreSQL on which it is based. This is necessary
to avoid any version conflict with existing drivers of PostgreSQL.
There are no modifications in the client/server connection protocol and psql.
However, a reference to Postgres-XC is added when calling version().
version() includes also a reference to the PostgreSQL version on which
Postgres-XC server is based.

In case additional commands dedicated to functionalities of Postgres-XC are
added in psql (simple command to show node list?) or whatever driver, it will
be necessary to extend a bit libpq with an additional API of the type
PQserverVersion to get XC dedicated version number. As there are no XC-dedicated
commands for the time being, this is not implemented yet.