Fix for deferred constraints
Basics for DEFERRED constraints were not working for 2 reasons.
1) There was a bug with the management of transaction-related parameters
(parameters that can be used only inside a transaction block like SET
CONSTRAINT and SET LOCAL). When a transaction parameter was set in a
transaction block, it was correctly sent down to remote nodes where
connections were already in hold, but if new connections were obtained
the transaction parameter on hold was sent to remote node before any
BEGIN query, resulting in having no effects.
2) If an implicit prepare failed, remote nodes were not correctly
rollbacked, resulting in remote connections keeping locks on prepared
relations. This could result by a stale state if for example a DDL was
involved in transaction.

Patch by Abbas Butt, review and some improvements by me.