Refactor transaction managemente code:
We track the nodes involved in global transaction as and when we send
BEGIN commands to the node. The nodes are tracked as either read-only
participant or read-write participants. This is important to later
decide whether do perform a 2PC or a simple commit on the remote
nodes. We right now rely on the planner information to (and the patch
does not change that much) to know if a transaction is doing any write
activity or not. This can be problematic in the presence of volatile
functions that can change the database state.

The decision to whether run a statement in auto-commit mode on the
data node or within a transaction block is also hard. We can run a
statement in auto-commit mode if and only if one data node is involved
in the transaction AND the statement will be sent to the data node
only once. Before this work, there were places where we were not
encapsulating statements in a transaction block possibly leading to
wrong results. After fixing this, there is a chance that we might see
some performance impact.

There are many cleanups: Here is a list, but probably not complete:

1. We now have only one member (transactionId) to track local or
global transaction identifier. Since we should always be using global
identifier in XC, this reduces the risk of wrong IDs being assigned
2. I removed the additional transaction state - TBLOCK_END_NOT_GTM. It
wasn't clear why we need this
3. 2PC is now hooked into PrepareTransaction/CommitTransaction.
Similarly, rollbacks and part commits are handled in AbortTransaction
4. The GTM transaction termination is handled in
Commit/AbortTransaction. That gives is a single entry point for
transaction management
5. Instead of various flags, we now just track if the local node is
involved in the transaction or not. That helps us decide whether to
perform 2PC on the local and remote node. This can further be enhanced
by looking at some other existing globals that PostgreSQL maintains
6. I have ripped off many other functions which are either not needed
now or probably needed for explicit 2PC. I will take another look at
them when I add back support for explicit 2PC