Support for SERIAL types
SERIAL columns in table use default values based on nextval of sequences
to auto-generate values. In vanilla Postgres, table creation with serial
column(s) is made with the following process:
1 - Create sequence(s)
2 - Create table
3 - Alter sequence(s) to change it as being owned by the column of table
previously created to manage object dependency.

In Postgres-XC, the sequence creation is made such as the object is created
on all the nodes, so a boolean flag associated to the serial process is
added to bypass sequence creation on remote nodes in case a sequence is created
within a serial process, and the query sent to remote nodes is the one given
by client application and it is sent only once when table is created on local
node.

Regression tests are all updated in consequence.