Add the support for Fast Query Shipping (FQS), a method to identify whether a query can be sent to the datanode/s as it is for evaluation and do so if deemed fit. In such cases, we create a plan with a single RemoteQuery node corresponding to the query and avoid the planning phase on coordinator.
A query tree walker analyses all the nodes in the query tree and finds out the
conditions under which the query is shippable and detects presence of
expressions which can not be evaluated on the datanode. It looks at the
relations involved in the query and deducts whether JOINs between these
relations can be evaluated on a single datanode.

In create_remotequery_plan() if the table is distributed, the nodelist is
reduced by looking at the qualifying conditions on the table. Function
create_remotejoin_plan is modified to use the FQS logic to decide the
shippability of JOIN.

Adds testcases xc_FQS and xc_FQS_join to test the fast query shipping
functionality and make it independent of cluster configuration.