Addition of initgtm: module for GTm/Proxy initialization
inigtm uses the infrastructure of initdb to provide the means to
initialize the data folder of a GTM or a GTM proxy. This is particularly
useful in case Postgres-XC binaries are installed in a folder where a user
cannot have access to it as, like initdb, it copies the sample configuration
files and makes them available for GTM instances.

initgtm has really a basic infrastructure now, but can be easily extended
to provide support for a cluster ID or settings related to GTM that need
to be decided at initialization.

Documentation is included.