GTM callback mechanism for sequence consistency on GTM
Addition of a mechanism to insure that sequence information
is kept consistent on GTM within a transaction block.
A callback to drop sequences on GTM is registered when sequences are
dropped or created. Then this callback is used when a transaction that created
sequence is aborted or when a transaction that dropped a sequence commits,
insuring that sequence information is properly rollbacked or committed on GTM.