Alternative regression output for test inherit
This is related to bug 3141171, where XC tries to perform joins with
objects Oids but it fails because OIDs are not consistent among nodes
in the cluster. This is a restriction of the XC current architecture.
Some additional queries, using a cast on regclass with relation name
are used to show the correct output in the test. This part of the test
has been updated in consequence.

There are still diffs in test inherit but it is not related to this
bug, so this regression still fails.