Correct remote DELETE planning for multiple relations
In case several relations were involved in a remote DELETE (rules,
join, etc.), remote query was not correctly generated because it
took into account not only its target entries, but also the ones
from other relations.
A reference of the original table is added in target entries and
this is compared to the relation whose remote DELETE is generated
to insure that all the operations are consistent.

This had consequences like bug 3512657, but also on the usage of
rules on DELETE.