Modify expected output files of timestamp.sql, timestamptz.sql to make them pass.
One change was to add ORDER BY clause in some queries that were left out.

Another change was to correct particular date values
that were displayed. The below commit
70d8a2c29e809c90f7caaded2b4f47b1cf18d01d in PG master branch has changed the
behaviour of FM suffix specification in to_char() function, so that the query:
select to_char('0097-02-16 17:32:01'::timestamp, 'FMYYY');
now returns 97, and before this commit it used to return 097.
And the expected change output in timestamp*.out has been updated in PG, but
did not get correctly merged in XC. In XC, the expected/timestamp.out still
had the old format values.

So the expected output files are now corrected accordingly in this commit.