Fix for bug 3485997: UPDATE fails on table with foreign key
This commit provides a fix for Coordinator trying to fetch a tuple
locally. This is not possible as all the data is localized on Datanodes.
This commit also enables reference integrity check on Datanodes.
This is thought as safe as XC has already some check mechanisms at constraint
creation to insure that constraint can be checked locally based on its dependant
table distributions.

Patch by Andrei Martsinchyk, I just fixed a bug with reference integrity
attempt check on Coordinator.