New GUC parameter to control 2PC usage for temporary objects
This new parameter, called enforce_two_phase_commit can be used to
enforce the use of 2PC on transactions involving temporary objects
or having ON COMMIT actions pending on them.
Enforcing the use of autocommit on several nodes that have performed
write operations at the same time might break consistency very easily
so the default value of enforce_two_phase_commit is on, but a superuser
can modify its value with SET on a private session. And its initial value
can be changed in postgresql.conf.
Regression tests have been updated in such a way that all the tests using
ON COMMIT actions or temporary objects have their sessions with this
parameter set to off, to it leads to a correct regression behavior.

The reason of introducing this parameter in Postgres-XC is originally
from PostgreSQL, where it is not possible to use two phase commit on
temporary objects. This might be or might not be implemented in Postgres,
so we need a mechanism to provide DBA some ways to protect data consistency
at all costs when he sets up his cluster as such functionality tunning
is application-dependent.