Addition of integer node identifier as part of tuple information
This mechanism permits to identify uniquely a tuple fetched from
Datanode on Coordinator by adding to the tuple information a field
called xc_node_id which is used as a unique node integer identifier.
This node ID is calculated at node creation and added in tuple information
as a kind of extension of ctid. It can be used in internal XC mechanisms.

This commit also includes a fix for round robin tables using remote update
and delete plans. This was an old issue that remained in the test xc_remote but
there was no way until now to identify the node origin of a tuple with a method
which is tuple-based.

Original feature is from Abbas.
The fix regarding round robin tables is from me. I also did some cleanup and
documentation addition on the new column of pgxc_node.