Fix for regression test prepared_xacts
Due to the new algorithm related to implicit/explicit 2PC handling,
local Coordinator where explicit PREPARE occurs is also prepared,
so change the output in consequence.
Also remove a transaction rollback message in test that was necessary
before new transaction handling because in case of a PREPARE transaction
failing its transaction was not properly rollbacked on remote nodes.