Fix for bug 3477909: COPY on table using nextval as default
To support nextval() in COPY, in general, execute a non-shippable default
expression on the coordinator, and add the default expression values in
the data row that is sent to the datanode. Earlier, because the value
was not evaluated and passed onto the datanode from coordinator, it was
getting evaluated on datanode, as a result of which nextval() used to fail.

Support this for both text and binary format.

This also fixes a bug in binary format COPY where it used to abort if there are
NULL values in the data.