Fix for node reduction in case of varchar data type.
Node reduction was not working in case of varchar data types and
the reason was that the function pgxc_find_distcol_expr was not
handling the RelabelType. This patch provide the proper handling.

Also the patch changes the expected output of the test case xc_distkey
to reflect changes in explain.