Fix for regression test array and simplify remote planner
xc_node_id has been introduced previously to insure that tuples fetched
from remote nodes are identified consistently in cluster. UPDATE and DELETE
remote planning has been simplified to use only xc_node_id and ctid in WHERE
clauses generated to be sure to interact with the good tuple.

This had consequences on arrays test where a check on a NULL array value
was done with = operator where it should have been an operator like IS NULL
or IS NOT NULL. Such an addition to WHERE clauses is unnecessary since
xc_node_id has been implemented.