Fix for regression test privileges
Related to issue 3520503, this test is failing because remote join query scan
in XC needs to more columns than necessary, sometimes including columns to
which users might not have access.
Reducing the target list on remote query scan is not that easy, so for the time
being this regression test is updated by being aware of this issue.