Remove additional output for test arrays
The distribution type of table used to check indexes on arrays is
changed to replication. Normally, a table with no supported distribution
types uses round robin as distribution but in the case of a constraint
being specified, round robin cannot be safely used to check constraints
on remote nodes and hash is enforced.
Related to feature request 3520520, this distribution type should be
changed to its initial value once global constraints are supported
in XC.