While generating create-table query for CTAS, include all reloptions as well.
The issue was, when WITH OIDS exists in create-table-as-select type of stmt, the
oids are not created on datanodes. This happens because the deparsed
create-table query generated for datanodes does not include WITH OIDS. So this
fix generates all reloptions that are supplied by the user into the deparsed
query.