Fix for bug 3489478: DML involving dropped attributes
When dropping a relation column, dropped columns were taken into account
in query generated for INSERT and UPDATE. Dropped columns are not taken into
account anymore in those queries and dropped parameters are casted in the
parameter list sent to remote nodes as int4, based on what Postgres returns
from the inner scan plan.

A new regression test called xc_alter_table is added to do some additional
check on those queries. The drop of the distribution column is now blocked
for safety, as this could cause damages on hash and Modulo tables.