Fix for bug ID : 3515461: DELETE query with non-pushable clauses does not work.
Two issues were fixed:

1. When the target list is not in the order of the table columns, WHERE clause
was not getting correctly applied. So, at the time of applying the quals, re-arrange
the slot values into an intermediate slot, so as to make the values ordered in line
with the order of the table attributes, and then apply the qual.

2. When WHERE clause contains subquery that refers to attributes in outer query,
the DELETE query fails, because the function pull_var_clause() that is used to
extract the qual vars into the target list does not peek into subquery. So replaced
this walker with a new walker that extracts vars from subquery as well.

More details about the fix are available in the bug comments.