Block creation of SQL functions containing utility statements
To support these functions we need to first implement a deparser
for all utility statements and then use the deparsed utility statement
while adding remote query node in ProcessUtility function.
Currently we are using "queryString" which can contain multiple
statements in case of SQL functions, some of which can be utility
and the last one has to be a select. Having such a combination in
a single remote query node creates problems while processing results.

Since that deparser is not there yet, we have to block creation of such SQL functions.

By setting the GUC check_function_bodies of OFF, one can by-pass the code
to block the creation of such functions. This patch makes the necessary
changes to make sure that such SQL functions do not produce any crash,
although they still won't run successfully.

The documentation is changed to let the users know of this limitation.

A few test cases are also added.