Block the use of INSERT-SELECT in case rows are inserted into a child by selecting from its parent.
This case had to be blocked because in this case XC was transforming the single INSERT-SELECT command into a multi-statement transaction block on the data nodes.

[STEP 1.]  START TRANSACTION ISOLATION LEVEL read committed READ WRITE
[STEP 2.]  SELECT col1,col2 FROM ONLY parent_table WHERE true
[STEP 3.]  INSERT INTO public.child_table (col1, col2) VALUES ($1, $2)
[STEP 4.]  (-- bind and execute here --)
[STEP 5.]  SELECT col1, col2 FROM ONLY child_table WHERE true
[STEP 6.]  INSERT INTO public.child_table (col1, col2) VALUES ($1, $2)
[STEP 7.]  (-- bind and execute here --)
[STEP 8.]  COMMIT TRANSACTION

The inserts of step 3 will be visible to select of step 5, hence we were getting duplicate rows in the child table.

To block we have added a check at the transformation stage.
Test cases are added to make sure block works fine for the case intended and does not block the rest of the cases.

Regressions showed a few extra failures which are fixed by the patch.