Improvement for make check
Improve the chances of interference with paralle installations by
fetching port values for each node and node poolers in private ranges.
There is still no way to check if GTM and pooler port values and
already in use like normal nodes where check is done by psql but
it reduces the chances of interference.