Block the use of system columns as MODULO/HASH columns
Those system columns are only used by local nodes, so it doesn't
make sense to use them as distribution key. This fix avoids a crash
that curiously only appeared when a CREATE TABLE using a system
column as distribution key was launched manually via psql, even if
a test is present in regressions.

Per report of Sutou Takayuki