Improve version number of psql by adding XC version number
The psql version of Postgres-XC has a couple of additions related to
commands exclusive to XC based on the new SQL commands introduced like
some TAB-recognition addition, so it makes sense to indicate the user
which is the psql version of XC used, in addition to the PostgreSQL
version number.
The startup output and the output of option --version are the only
parts modified.