Fix for test money for non-English based environments
An issue with the test case money happened in non-English environments
because the GUC parameter lc_monetary was not taken into account on
remote Datanodes.
So like time and interval parameters, it is passed to pooler who is in
charge to use the correct value at connection creation.
lc_monetary does not accept lower case value, so its value obtained
directly from GUC API is used.
This issue has been reproduced in a Japanese environment.

Per report of Sutou Takayuki